to long

+++ CardSharp/GameComponents/Desk.cs
@@ -271,7 +271,7 @@ namespace CardSharp
             {
                 AddMessageLine("SDDC duel done.");
 
-                int result = 0;
+                long result = 0;
                 switch (player.Type) {
                     case PlayerType.Farmer:
                         AddMessageLine("Winners are farmers.");
@@ -314,7 +314,7 @@ namespace CardSharp
             FinishGame(true);
         }
 
-        private int SaveSddc(IEnumerable<Player> winners, IEnumerable<Player> losers)
+        private long SaveSddc(IEnumerable<Player> winners, IEnumerable<Player> losers)
         {
             var winnersConfig = winners.Select(PlayerConfig.GetConfig).ToList();
             var losersConfig = losers.Select(PlayerConfig.GetConfig).ToList();
@@ -328,7 +328,7 @@ namespace CardSharp
             return score;
         }
 
-        private void SaveScore(PlayerConfig p, int value)
+        private void SaveScore(PlayerConfig p, long value)
         {
             var playerConf = p;
             playerConf.Point = value;
