+++ CardSharp/GameComponents/IPlayerConfig.cs
@@ -5,7 +5,7 @@ namespace CardSharp.GameComponents
     public interface IPlayerConfig
     {
         DateTime LastTime { get; set; }
-        int Point { get; set; }
+        long Point { get; set; }
         string PlayerID { get; }
         bool IsAdmin { get; }
     }
